+++ CatEngine/src/Platform/Windows/WindowsWindow.cpp
@@ -190,7 +190,7 @@ namespace CatEngine {
 		// Mouse Position Callback
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xpos, double ypos) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseMovedEvent event((unsigned int)xpos, (unsigned int)ypos);
+			MouseMovedEvent event((const float)(unsigned int)xpos, (const float)(unsigned int)ypos);
 			data.EventCallback(event);
 		});
 	}
